lib/python/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "satos_payload"
authors = [
  {name = "Antaris, Inc", email = "opensource@antaris.space"},
]
description = "Library used to develop SatOS(TM) payloads"
license = { text = "Apache-2.0" }
requires-python = ">=3.7"
dependencies = [
  "importlib-metadata; python_version <\"3.8\"",
  "grpcio ~=1.0",
  "grpcio-tools",
  "protobuf ~=4.0",
  "pylibftdi",
]
#version = "1.1.1"
dynamic = [
  "version"
]

[tool.setuptools.dynamic]
version = { attr = "satos_payload.gen.antaris_sdk_version.VERSION" }
